+++ modules/Custom/RequestsMatrix.php
@@ -4,7 +4,7 @@
 r.COURSE_ID,cp.COURSE_PERIOD_ID,
 c.TITLE AS COURSE_TITLE,cp.PERIOD_ID,
 (cp.TOTAL_SEATS-cp.FILLED_SEATS) AS OPEN_SEATS,s.STUDENT_ID AS SCHEDULED
-FROM SCHEDULE_REQUESTS r,
+FROM schedule_requests r,
 courses c,SCHOOL_PERIODS sp,
 course_periods cp LEFT OUTER JOIN SCHEDULE s ON
 (s.COURSE_PERIOD_ID=cp.COURSE_PERIOD_ID AND s.STUDENT_ID='".UserStudentID()."')
@@ -18,7 +18,7 @@ ORDER BY ".db_case(array('s.STUDENT_ID',"''","NULL",'sp.SORT_ORDER'))."
 $requests_RET = DBGet( "SELECT r.COURSE_ID AS CRS,r.COURSE_ID,cp.COURSE_PERIOD_ID,
 c.TITLE AS COURSE_TITLE,cp.PERIOD_ID,
 (cp.TOTAL_SEATS-cp.FILLED_SEATS) AS OPEN_SEATS,s.STUDENT_ID AS SCHEDULED
-FROM SCHEDULE_REQUESTS r,courses c,SCHOOL_PERIODS sp,course_period_school_periods cpsp,course_periods cp
+FROM schedule_requests r,courses c,SCHOOL_PERIODS sp,course_period_school_periods cpsp,course_periods cp
 LEFT OUTER JOIN SCHEDULE s ON (s.COURSE_PERIOD_ID=cp.COURSE_PERIOD_ID AND s.STUDENT_ID='" . UserStudentID() . "')
 WHERE cp.COURSE_PERIOD_ID=cpsp.COURSE_PERIOD_ID
 AND r.SYEAR='" . UserSyear() . "'