+++ modules/Scheduling/new_Requests.php
@@ -22,7 +22,7 @@ if ( $_REQUEST['modfunc'] == 'verify' )
 		WHERE SCHOOL_ID='" . UserSchool() . "'
 		AND SYEAR='" . UserSyear() . "'", [], [ 'COURSE_ID' ] );
 
-	DBQuery( "DELETE FROM SCHEDULE_REQUESTS
+	DBQuery( "DELETE FROM schedule_requests
 		WHERE STUDENT_ID='" . UserStudentID() . "'
 		AND SYEAR='" . UserSyear() . "'" );
 
@@ -39,7 +39,7 @@ if ( $_REQUEST['modfunc'] == 'verify' )
 				continue;
 			}
 
-			$sql = "INSERT INTO SCHEDULE_REQUESTS (SYEAR,SCHOOL_ID,STUDENT_ID,SUBJECT_ID,COURSE_ID,MARKING_PERIOD_ID,WITH_TEACHER_ID,NOT_TEACHER_ID,WITH_PERIOD_ID,NOT_PERIOD_ID)
+			$sql = "INSERT INTO schedule_requests (SYEAR,SCHOOL_ID,STUDENT_ID,SUBJECT_ID,COURSE_ID,MARKING_PERIOD_ID,WITH_TEACHER_ID,NOT_TEACHER_ID,WITH_PERIOD_ID,NOT_PERIOD_ID)
 			VALUES('" . UserSyear() . "','" .
 			UserSchool() . "','" . UserStudentID() . "','" . $courses_RET[$course][1]['SUBJECT_ID'] . "','" .
 			$course . "',NULL,'" . $_REQUEST['with_teacher'][$subject][$i] . "','" .
@@ -62,7 +62,7 @@ if ( $_REQUEST['modfunc'] == 'choose' )
 {
 	$functions = [ 'WITH_PERIOD_ID' => '_makeWithSelects', 'NOT_PERIOD_ID' => '_makeWithoutSelects' ];
 	$requests_RET = DBGet( "SELECT sr.COURSE_ID,c.COURSE_TITLE,sr.WITH_PERIOD_ID,sr.NOT_PERIOD_ID,sr.WITH_TEACHER_ID,sr.NOT_TEACHER_ID
-	FROM SCHEDULE_REQUESTS sr,courses c
+	FROM schedule_requests sr,courses c
 	WHERE sr.SYEAR='" . UserSyear() . "'
 	AND sr.STUDENT_ID='" . UserStudentID() . "'
 	AND sr.COURSE_ID=c.COURSE_ID", $functions );