+++ CHANGES.md
@@ -66,7 +66,7 @@ Changes in 9.1.1
 - Fix Color Input was hidden in ReportCardComments.php
 - Fix use Course ID in session in MassRequests.php
 - Fix SQL error primary key exists on table food_service_staff_accounts in Rollover.php
-- Fix SQL error foreign key exists on tables gradebook_assignments,gradebook_assignment_types,SCHEDULE_REQUESTS in Rollover.php
+- Fix SQL error foreign key exists on tables gradebook_assignments,gradebook_assignment_types,schedule_requests in Rollover.php
 - Fix save State input value in Registration.fnc.php
 - Fix SchoolInfo() on user School Year update in School.php